feat(ts): ts改造
