--- es/const/ws.js
@@ -1,8 +1,0 @@
-/**
- * websocket消息类型
- * @type {{RELOAD: string, OFFLINE: string}}
- */
-export const WEBSOCKET_ACTION = {
-    RELOAD: 'reload',
-    OFFLINE: 'offline',
-};
