--- es/utils/index.js
@@ -1,2 +1,0 @@
-export * from './auth';
-export * from './util';
