--- lib/const/ws.d.ts
@@ -1,8 +1,0 @@
-/**
- * websocket消息类型
- * @type {{RELOAD: string, OFFLINE: string}}
- */
-export declare const WEBSOCKET_ACTION: {
-    RELOAD: string;
-    OFFLINE: string;
-};
