--- lib/const/ws.js
@@ -1,11 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.WEBSOCKET_ACTION = void 0;
-/**
- * websocket消息类型
- * @type {{RELOAD: string, OFFLINE: string}}
- */
-exports.WEBSOCKET_ACTION = {
-    RELOAD: 'reload',
-    OFFLINE: 'offline',
-};
