+++ package.json
@@ -2,26 +2,27 @@
   "name": "node-agile-client",
   "version": "0.0.2",
   "description": "AgileConfig的node客户端",
-  "main": "./lib/index.js",
-  "module": "./es/index.js",
-  "jsnext:main": "./es/index.js",
-  "types": "./lib/index.d.ts",
-  "files": [
-    "es",
-    "lib"
-  ],
+  "main": "./cjs/index.js",
+  "module": "./esm5/index.js",
+  "typings": "./types/index.d.ts",
   "scripts": {
-    "build:commonjs": "rimraf lib && tsc --target es5 --outDir lib --module commonjs --declaration",
-    "build:es": "rimraf es && tsc --target es6 --outDir es --module es6",
-    "build": "npm run build:commonjs && npm run build:es",
-    "ts": "ts-node src/",
-    "prepublishOnly": "npm run build"
+    "build:clean": "rimraf build",
+    "build": "npm run build:clean && npm run build:cjs && npm run build:esm5 && npm run build:types && npm run build:cp",
+    "build:esm5": "tsc --project tsconfig.esm5.json",
+    "build:cjs": "tsc --project tsconfig.cjs.json",
+    "build:types": "tsc --project tsconfig.types.json",
+    "build:cp": "cp LICENSE build/LICENSE && cp README.md build/README.md && cp package.json build/package.json"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/LetTTGACO/node-agile-client.git"
   },
-  "keywords": [],
+  "tags": [
+    "agile"
+  ],
+  "keywords": [
+    "配置中心"
+  ],
   "author": "",
   "license": "ISC",
   "bugs": {
