+++ tsconfig.esm5.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "ES2015",
+    "target": "ES5",
+    "outDir": "build/esm5",
+  },
+}
